% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_partitioned_data.R
\name{combine_partitioned_data}
\alias{combine_partitioned_data}
\title{Combine partitioned ARTIS data files into a single Parquet file}
\usage{
combine_partitioned_data(
  search_dir,
  outdir,
  data_type = c("trade", "consumption"),
  estimate_data_type = c("midpoint", "max", "min"),
  artis_version = "v1.0.0",
  prod_data_type = prod_data_type,
  file_type = c("qs2", "qdata", "csv", "parquet"),
  date = format(Sys.Date(), "\%Y-\%m-\%d"),
  search_pattern,
  custom_timeseries = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{search_dir}{Directory where partitioned files are stored.}

\item{outdir}{Output directory for the combined parquet file.}

\item{data_type}{Data type: "trade" (snet), "consumption", etc.}

\item{estimate_data_type}{Which estimate: e.g., "midpoint", "max", "min".}

\item{artis_version}{Model version string, e.g. "v1.2.0".}

\item{prod_data_type}{Production model configuration used for file naming only.}

\item{file_type}{Input file format. Options: "qs" (default), "csv", "parquet".}

\item{date}{Date string for output file naming (default: today's date).}

\item{search_pattern}{Regex to match the desired files.}

\item{custom_timeseries}{Logical. If TRUE, adds "custom_ts" to output filename. FIXIT: add this functionality.}

\item{verbose}{Logical. Print file names as they are added.}
}
\description{
Combine partitioned ARTIS data files into a single Parquet file
}
